gitignore: Add .gitignore to the project
.gtignore is a simple file which includes the names of folders and files.
These files and folders will be excluded when someone with use "git add ." or "git add -A".
In simple words, these files and folders will not be staged for commit ever as these files are large and dependencies are already added to package.json.
For more info: https://help.github.com/articles/ignoring-files/

Signed-off-by: Prasoon Mehra <prasoonmehra14@gmail.com>
Change-Id: I56ca46bbd22689308d051af15027feab23b6c2bb